Move the ProgressHeader and HeaderIconButton component to folder 'Share'
